requirements.txt
fastapi>=0.100.0
uvicorn>=0.23.0
tushare>=1.4.0
psycopg2-binary>=2.9.0
sqlalchemy[asyncio]>=2.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
pyyaml>=6.0
python-dateutil>=2.8.0
alembic>=1.12.0
asyncpg>=0.29.0

